embed-documentation/requirements.txt
openai>=1.6.0
python-dotenv

